index.js: Fix regression breaking automatic reselection of destination user upon in-thread DM reply
This issue arose only when sending a threaded reply to a user in a *different* workspace.  It was
possible to work around the issue by using the slash command to manually select the user.

The problem was the use of the wrong user (local instead of remote) when checking whether they had
since been removed from the remote side of the bridged channel.  As part of this fix, we had to add
a new, currently O(n), reverse-lookup feature to find the sender by their DM channel; this is
necessary because they might no longer be DM'ing the same person or the server might have restarted.