index.js: Populate im->uid mappings during bootstrapping and channel join
This fixes a bug where the uid->im reverse lookup would fail when posting a threaded response to a
DM after a server response not followed by explicit selection of the same destination user.

The direct-message caches are not used as a source of truth for membership, so there is no need to
remove entries when a member leaves a channel.